+++ Manager/ManagerBioquimico.cpp
@@ -7,8 +7,8 @@ ManagerBioquimico::ManagerBioquimico(const std::string& nombreArchivo): Manager<
 //ORDENAR POR APELLIDO - BURBUJA
 std::vector<Bioquimico> ManagerBioquimico::ordenarPorApellido() {
     std::vector<Bioquimico> lista = leerTodos();
-    for (int i = 0; i < lista.size() - 1; i++) {
-        for (int j = i + 1; j < lista.size(); j++) {
+    for (int i = 0; i < (int)lista.size() - 1; i++) {
+        for (int j = i + 1; j < (int)lista.size(); j++) {
             if (lista[i].getApellido() > lista[j].getApellido()) {
                 Bioquimico aux = lista[i];
                 lista[i] = lista[j];
@@ -22,8 +22,8 @@ std::vector<Bioquimico> ManagerBioquimico::ordenarPorApellido() {
 //ORDENAR POR LEGAJO - BURBUJA
 std::vector<Bioquimico> ManagerBioquimico::ordenarPorLegajo() {
     std::vector<Bioquimico> lista = leerTodos();
-    for (int i = 0; i < lista.size() - 1; i++) {
-        for (int j = i + 1; j < lista.size(); j++) {
+    for (int i = 0; i < (int)lista.size() - 1; i++) {
+        for (int j = i + 1; j < (int)lista.size(); j++) {
             if (lista[i].getMatricula() > lista[j].getMatricula()) {
                 Bioquimico aux = lista[i];
                 lista[i] = lista[j];
