+++ Manager/ManagerCategoria.cpp
@@ -7,7 +7,7 @@ ManagerCategoria::ManagerCategoria(const std::string& nombreArchivo) : Manager<C
 //BUSCAR POR NOMBRE
 Categoria ManagerCategoria::buscarPorNombre(std::string nombre, bool& encontro) {
     std::vector<Categoria> lista = leerTodos();
-    for (int i = 0; i < lista.size(); i++) {
+    for (int i = 0; i < (int)lista.size(); i++) {
         if (lista[i].getNombre() == nombre){
             encontro = true;
             return lista[i];
