+++ Manager/ManagerObraSocial.cpp
@@ -4,7 +4,7 @@ ManagerObraSocial::ManagerObraSocial(const std::string& nombreArchivo):Manager<O
 
 ObraSocial ManagerObraSocial::buscarPorNombre(std::string& nombre){
     std::vector<ObraSocial> obrasSociales = leerTodos();
-    for(int i = 0;i<obrasSociales.size();i++){
+    for(int i = 0; i < (int)obrasSociales.size();i++){
         if(obrasSociales[i].getNombre() == nombre){
             return obrasSociales[i];
         }
