+++ Manager/ManagerPaciente.cpp
@@ -10,8 +10,8 @@ std::vector<Paciente> ManagerPaciente::ordenarPorApellido() {
     std::vector<Paciente> lista = leerTodos();
 
     //Ordenamiento burbuja -> Comparamos un registro con otro y vemos cual es mas grande
-    for (int i = 0; i < lista.size() - 1; i++) {
-        for (int j = i + 1; j < lista.size(); j++) {
+    for (int i = 0; i < (int)lista.size() - 1; i++) {
+        for (int j = i + 1; j < (int)lista.size(); j++) {
             std::string ap1 = lista[i].getApellido();
             std::string ap2 = lista[j].getApellido();
             if (ap1 > ap2) {
@@ -30,7 +30,7 @@ std::vector<Paciente> ManagerPaciente::ordenarPorApellido() {
 Paciente ManagerPaciente::buscarPorDNI(int dni) {
     std::vector<Paciente> lista = leerTodos();
 
-    for (int i = 0; i < lista.size(); i++) {
+    for (int i = 0; i < (int)lista.size(); i++) {
         if (lista[i].getDNI() == dni) {
             return lista[i];
         }
@@ -44,7 +44,7 @@ std::vector<Paciente> ManagerPaciente::buscarPorApellido(std::string apellido) {
     std::vector<Paciente> lista = leerTodos();
     std::vector<Paciente> resultados;
 
-    for (int i = 0; i < lista.size(); i++) {
+    for (int i = 0; i < (int)lista.size(); i++) {
         if (lista[i].getApellido() == apellido) {
             resultados.push_back(lista[i]); //Metemos el registro que encontramos al final del vector de resultado
         }
@@ -58,7 +58,7 @@ std::vector<Paciente> ManagerPaciente::buscarPorObraSocial(int idObraSocial) {
     std::vector<Paciente> lista = leerTodos();
     std::vector<Paciente> resultados;
 
-    for (int i = 0; i < lista.size(); i++) {
+    for (int i = 0; i < (int)lista.size(); i++) {
         if (lista[i].getCodigoObraSocial() == idObraSocial) {
             resultados.push_back(lista[i]); //Igual que arriba nada mas que por obra social -> reutilizar logica para toda busqueda con int
         }
