+++ Manager/ManagerTurno.cpp
@@ -7,8 +7,8 @@ ManagerTurno::ManagerTurno(const std::string& nombreArchivo) : Manager<Turno>(no
 // ORDENAR POR FECHA
 std::vector<Turno> ManagerTurno::ordenarPorFecha() {
     std::vector<Turno> lista = leerTodos();
-    for (int i = 0; i < lista.size() - 1; i++) {
-        for (int j = i + 1; j < lista.size(); j++) {
+    for (int i = 0; i < (int)lista.size() - 1; i++) {
+        for (int j = i + 1; j < (int)lista.size(); j++) {
             if (lista[i].getFechaTurno().esAnterior(lista[j].getFechaTurno())) {
                 Turno aux = lista[i];
                 lista[i] = lista[j];
@@ -25,7 +25,7 @@ std::vector<Turno> ManagerTurno::buscarPorFecha(int dia, int mes, int anio) {
     std::vector<Turno> lista = leerTodos();
     std::vector<Turno> resultado;
 
-    for (int i = 0; i < lista.size(); i++) {
+    for (int i = 0; i < (int)lista.size(); i++) {
         FechaHora fechaHora = lista[i].getFechaTurno();
         if (fechaHora.getDia() == dia && fechaHora.getMes() == mes && fechaHora.getAnio() == anio) {
             resultado.push_back(lista[i]);
@@ -39,7 +39,7 @@ std::vector<Turno> ManagerTurno::buscarPorPaciente(int idPaciente) {
     std::vector<Turno> lista = leerTodos();
     std::vector<Turno> resultado;
 
-    for (int i = 0; i < lista.size(); i++) {
+    for (int i = 0; i < (int)lista.size(); i++) {
         if (lista[i].getIDPaciente() == idPaciente) {
             resultado.push_back(lista[i]);
         }
@@ -52,7 +52,7 @@ std::vector<Turno> ManagerTurno::buscarPorBioquimico(int idBioquimico) {
     std::vector<Turno> lista = leerTodos();
     std::vector<Turno> resultado;
 
-    for (int i = 0; i < lista.size(); i++) {
+    for (int i = 0; i < (int)lista.size(); i++) {
         if (lista[i].getIDBioquimico() == idBioquimico) {
             resultado.push_back(lista[i]);
         }
@@ -65,7 +65,7 @@ std::vector<Turno> ManagerTurno::buscarPorAsistencia(){
     std::vector<Turno> lista = leerTodos();
     std::vector<Turno> resultado;
 
-    for(int i = 0; i < lista.size(); i++){
+    for(int i = 0; i < (int)lista.size(); i++){
         if(lista[i].getAsistio() == true){
             resultado.push_back(lista[i]);
         }
